+++ src/runtime.js
@@ -126,11 +126,15 @@ function makeKeywordArgs(obj) {
     return obj;
 }
 
+function isKeywordArgs(obj) {
+  return obj && Object.prototype.hasOwnProperty.call(obj, '__keywords');
+}
+
 function getKeywordArgs(args) {
     var len = args.length;
     if(len) {
         var lastArg = args[len - 1];
-        if(lastArg && lastArg.hasOwnProperty('__keywords')) {
+        if(isKeywordArgs(lastArg)) {
             return lastArg;
         }
     }
@@ -144,7 +148,7 @@ function numArgs(args) {
     }
 
     var lastArg = args[len - 1];
-    if(lastArg && lastArg.hasOwnProperty('__keywords')) {
+    if(isKeywordArgs(lastArg)) {
         return len - 1;
     }
     else {